pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "This project contains a module that defines the differences between the two files. File formats: -JSON; -YAML; -YML. Output formats: -Stylish; -Plain; -JSON."
authors = ["Valerii-Denisov <klerk919191@gmail.com>"]
packages = [{ include = "gendiff"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: Free for non-commercial use",
    "Operating System :: POSIX",
    "Programming Language :: Python",
]
repository = "https://github.com/Valerii-Denisov/python-project-lvl2"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
wemake-python-styleguide = "^0.16.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
